Move functions from course lib to address circular dependency
