+++ apps/prairielearn/src/lib/course.sql
@@ -17,10 +17,3 @@ FROM
 WHERE
   cp.course_role = 'Owner'
   AND cp.course_id = $course_id;
-
--- BLOCK update_course_commit_hash
-UPDATE pl_courses
-SET
-  commit_hash = $commit_hash
-WHERE
-  id = $course_id;
