+++ apps/prairielearn/src/lib/course.ts
@@ -1,16 +1,20 @@
-import { callbackify, promisify } from 'util';
-import { exec } from 'child_process';
+import { callbackify } from 'util';
 import * as fs from 'fs-extra';
 import * as namedLocks from '@prairielearn/named-locks';
 import * as sqldb from '@prairielearn/postgres';
-import * as error from '@prairielearn/error';
 
 import { createServerJob } from './server-jobs';
 import { config } from './config';
 import * as chunks from './chunks';
 import { syncDiskToSqlWithLock } from '../sync/syncFromDisk';
 import { IdSchema, User, UserSchema } from './db-types';
-import { selectCourseById } from '../models/course';
+import {
+  getCommitHash,
+  getLockNameForCoursePath,
+  getOrUpdateCourseCommitHash,
+  selectCourseById,
+  updateCourseCommitHash,
+} from '../models/course';
 
 const sql = sqldb.loadSqlEquiv(__filename);
 
@@ -43,58 +47,6 @@ export const checkBelongs = callbackify(checkBelongsAsync);
  */
 export async function getCourseOwners(course_id: string): Promise<User[]> {
   return await sqldb.queryRows(sql.select_owners, { course_id }, UserSchema);
-}
-
-export function getLockNameForCoursePath(coursePath: string): string {
-  return `coursedir:${coursePath}`;
-}
-
-export async function getCommitHash(coursePath: string): Promise<string> {
-  try {
-    const { stdout } = await promisify(exec)('git rev-parse HEAD', {
-      cwd: coursePath,
-      env: process.env,
-    });
-    return stdout.trim();
-  } catch (err) {
-    throw error.makeWithData(`Could not get git status; exited with code ${err.code}`, {
-      stdout: err.stdout,
-      stderr: err.stderr,
-    });
-  }
-}
-
-/**
- * Loads the current commit hash from disk and stores it in the database. This
- * will also add the `commit_hash` property to the given course object.
- */
-export async function updateCourseCommitHash(course: {
-  id: string;
-  path: string;
-}): Promise<string> {
-  const hash = await getCommitHash(course.path);
-  await sqldb.queryAsync(sql.update_course_commit_hash, {
-    course_id: course.id,
-    commit_hash: hash,
-  });
-  return hash;
-}
-
-/**
- * If the provided course object contains a commit hash, that will be used;
- * otherwise, the commit hash will be loaded from disk and stored in the
- * database.
- *
- * This should only ever really need to happen at max once per course - in the
- * future, the commit hash will already be in the course object and will be
- * updated during course sync.
- */
-export async function getOrUpdateCourseCommitHash(course: {
-  id: string;
-  path: string;
-  commit_hash?: string | null;
-}): Promise<string> {
-  return course.commit_hash ?? (await updateCourseCommitHash(course));
 }
 
 export async function pullAndUpdateCourse({
