+++ apps/prairielearn/src/lib/github.js
@@ -6,7 +6,7 @@ const ERR = require('async-stacktrace');
 const { config } = require('./config');
 const { logger } = require('@prairielearn/logger');
 const serverJobs = require('./server-jobs-legacy');
-const { updateCourseCommitHash } = require('./course');
+const { updateCourseCommitHash } = require('../models/course');
 const syncFromDisk = require('../sync/syncFromDisk');
 const opsbot = require('./opsbot');
 const chunks = require('./chunks');
