+++ apps/prairielearn/src/models/course.sql
@@ -5,3 +5,10 @@ FROM
   pl_courses
 where
   id = $course_id;
+
+-- BLOCK update_course_commit_hash
+UPDATE pl_courses
+SET
+  commit_hash = $commit_hash
+WHERE
+  id = $course_id;
