+++ apps/prairielearn/src/question-servers/freeform.js
@@ -18,7 +18,7 @@ const { logger } = require('@prairielearn/logger');
 const { withCodeCaller, FunctionMissingError } = require('../lib/code-caller');
 const jsonLoad = require('../lib/json-load');
 const cache = require('../lib/cache');
-const { getOrUpdateCourseCommitHash } = require('../lib/course');
+const { getOrUpdateCourseCommitHash } = require('../models/course');
 const markdown = require('../lib/markdown');
 const chunks = require('../lib/chunks');
 const assets = require('../lib/assets');
